requirements.txt
psutil==5.9.0
websocket-client==1.3.2
pytz==2022.1
alpaca_trade_api
mysql-connector-python~=8.0.27
pandas==1.1.5
yfinance~=0.1.67
yahoo-fin
ta==0.10.1
numpy~=1.21.4
pytest==7.1.2
python-dotenv==0.20.0
setuptools==62.1.0
pandas-ta
PyDispatcher~=2.0.5
urllib3~=1.26.9
yarl~=1.7.2